+++ tests/Attribute/Parameter/TrimTest.php
@@ -38,6 +38,7 @@ final class TrimTest extends TestCase
             TestHelper::getFirstParameter(static fn(?string $a) => null),
             Result::success($value),
             new ArrayData(),
+            new Hydrator(),
         );
 
         $result = $resolver->getParameterValue($attribute, $context);