+++ tests/Support/Classes/Chart/ChartSet.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Hydrator\Tests\Support\Classes\Chart;
+
+use Yiisoft\Hydrator\Attribute\Parameter\Collection;
+
+final class ChartSet
+{
+    public function __construct(
+        #[Collection(Chart::class)]
+        private array $charts = [],
+    ) {
+    }
+}