+++ tests/Support/Classes/Chart/Point.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Hydrator\Tests\Support\Classes\Chart;
+
+final class Point
+{
+    public function __construct(
+        private Coordinates $coordinates,
+        private array $rgb,
+    ) {
+    }
+}