+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.3.0 under development
 
+- New #49: Add `Collection` PHP attribute (@arogachev)
+- New #49: Add hydrator dependency and `withHydrator()` method to `ParameterAttributesHandler` (@arogachev)
+- New #49: Add hydrator dependency and `getHydrator()` method to `ParameterAttributeResolveContext` (@arogachev)
 - Enh #85: Allow to hydrate non-initialized readonly properties (@vjik)
 
 ## 1.2.0 April 03, 2024